Authors & Default Slider
Authors have been added to the Credits Page.
Public variable Slider called volumeSlider has been created (and connected to the Slider from Settings in Editor).
The value position of the Slider is now appropriate to the actual volume value (before, it was always resetting. Now it is saving properly)
After pressing the Revert Button the value on the slider will be 0.7f, this will also affect the music playing at the moment.